Lots of Groundwork Gave up on writting a log. Cout and the like are probably enough for now.
Began work on the rpg stuff I'm gonna need. Made an Attribute class
that stores the current value of an attribute and binds it to a specific
range of values. The bounds can be changed and reset to a predetermined
value. The predetermined values can also be adjusted. I overloaded most
of the operator to interact with ints, they don't throw errors. The
setters will throw errors though. The class guarantees the stored value
is within the bounds and the upper and lower bounds are ordered.